mostly organizing test code + wrote some tests for stripTags
