--- pkg/skaffold/initializer/util.go
@@ -1,27 +1,0 @@
-/*
-Copyright 2019 The Skaffold Authors
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package initializer
-
-import "github.com/GoogleContainerTools/skaffold/pkg/skaffold/schema"
-
-// isSkaffoldConfig is for determining if a file is skaffold config file.
-func isSkaffoldConfig(file string) bool {
-	if config, err := schema.ParseConfig(file, false); err == nil && config != nil {
-		return true
-	}
-	return false
-}
